test(e2e): address provider review findings
