+++ tests/e2e/provider_operation_cases.py
@@ -3,8 +3,8 @@
 from provider_operation_github_issue_cases import (
     GITHUB_ISSUE_PROVIDER_OPERATION_CASES,
 )
-from provider_operation_github_gap_cases import (
-    GITHUB_GAP_PROVIDER_OPERATION_CASES,
+from provider_operation_github_actions_cases import (
+    GITHUB_ACTIONS_PROVIDER_OPERATION_CASES,
 )
 from provider_operation_github_pull_cases import (
     GITHUB_PULL_PROVIDER_OPERATION_CASES,
@@ -28,8 +28,8 @@ PROVIDER_OPERATION_CASES = (
     *GOOGLE_DRIVE_PROVIDER_OPERATION_CASES,
     *GOOGLE_SHEETS_PROVIDER_OPERATION_CASES,
     *GOOGLE_SLIDES_PROVIDER_OPERATION_CASES,
+    *GITHUB_ACTIONS_PROVIDER_OPERATION_CASES,
     *GITHUB_ISSUE_PROVIDER_OPERATION_CASES,
     *GITHUB_PULL_PROVIDER_OPERATION_CASES,
     *GITHUB_REPO_PROVIDER_OPERATION_CASES,
-    *GITHUB_GAP_PROVIDER_OPERATION_CASES,
 )
